Fix pointer events parsing with Chromium

+++ src/Uno.UI/UI/Xaml/Input/PointerRoutedEventArgs.wasm.cs
@@ -13,13 +13,13 @@ namespace Windows.UI.Xaml.Input
 {
 	partial class PointerRoutedEventArgs
 	{
-		private readonly ulong _timestamp;
+		private readonly double _timestamp;
 		private readonly Point _absolutePosition;
 		private readonly VirtualKey _button;
 		private readonly PointerUpdateKind _updateKind;
 
 		internal PointerRoutedEventArgs(
-			ulong timestamp,
+			double timestamp,
 			uint pointerId,
 			PointerDeviceType pointerType,
 			Point absolutePosition,
@@ -77,17 +77,17 @@ namespace Windows.UI.Xaml.Input
 		#region Misc static helpers
 		private static ulong? _bootTime;
 
-		private static ulong ToTimeStamp(ulong timestamp)
+		private static ulong ToTimeStamp(double timestamp)
 		{
 			if (!_bootTime.HasValue)
 			{
-				_bootTime = ulong.Parse(WebAssemblyRuntime.InvokeJS("Date.now() - performance.now()")) * TimeSpan.TicksPerMillisecond;
+				_bootTime = (ulong) (double.Parse(WebAssemblyRuntime.InvokeJS("Date.now() - performance.now()")) * TimeSpan.TicksPerMillisecond);
 			}
 
-			return _bootTime.Value + (timestamp * TimeSpan.TicksPerMillisecond);
+			return _bootTime.Value + (ulong)(timestamp * TimeSpan.TicksPerMillisecond);
 		}
 
-		private static uint ToFrameId(ulong timestamp)
+		private static uint ToFrameId(double timestamp)
 		{
 			// Known limitation: After 49 days, we will overflow the uint and frame IDs will restart at 0.
 			return (uint)(timestamp % uint.MaxValue);
